requirements.txt
asgiref==3.11.0
Django==6.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
promise==2.3
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.4
text-unidecode==1.3
typing_extensions==4.15.0
django-cors-headers==4.3.1
Pillow==10.2.0
